PimForceReact/backend/requirements.txt
# Django and DRF
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3

# Database
psycopg2-binary==2.9.7
psycopg==3.1.18
dj-database-url==2.1.0

# Authentication
djangorestframework-simplejwt==5.3.0

# API Documentation
drf-spectacular==0.26.5

# Background Tasks
celery==5.3.4
redis==5.0.1

# Environment and Settings
python-decouple==3.8
django-environ==0.11.2

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Development Tools
django-debug-toolbar==4.2.0

# Code Quality
black==23.11.0
isort==5.12.0
mypy==1.7.1
flake8==6.1.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Utilities
Pillow==10.1.0
python-dateutil==2.8.2
requests==2.31.0

